requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
dialogflow==1.0.0
feedparser==5.2.1
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.21.0
google-auth==1.18.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
html5lib==1.1
idna==2.9
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
parso==0.7.0
pickleshare==0.7.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
python-dateutil==2.8.1
python-telegram-bot==12.7
pytz==2019.1
pyzmq==19.0.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
